chore: sync

+++ trixie/install.sh
@@ -1,38 +1,42 @@
 #! /bin/bash
 
-venv_name=""
-if [ $# -eq 0 ]; then
-    venv_name=".venv"
-elif [ $# -eq 1 ]; then
-    venv_name=$1
-else
-    echo "Error: Invalid arguments."
-    exit 1
+read -p "Is Python environment already set up? [y/N]: " answer
+
+if [[ ! "$answer" =~ ^[yY]$ ]]; then
+    # Confirm Python version & paths
+    module load python/3.12.1-gcc-11.3.1-7tuhjhr
+    module list 
+    which python
+    which pip
+    python --version
+
+    read -p "Name of Python virtual environment: " venv_name
+    mkdir -p $HOME/venvs
+    if [ -d "$HOME/venvs/$venv_name" ]; then
+        rm -rf $HOME/venvs/$venv_name
+    fi
+    python3 -m venv $HOME/venvs/$venv_name
+    ln -s $HOME/venvs/$venv_name $venv_name
+    source $venv_name/bin/activate
+
+    pip install -U pip
+    pip install inspect_ai \
+                git+https://github.com/UKGovernmentBEIS/inspect_evals \
+                openai \
+                # vllm \
+                # mistral_common # Mistral-Large-Instruct-2411
+    # pip freeze > requirements.txt
+    deactivate
+    echo -e "\nSuccessfully installed Python packages to '$venv_name'."
 fi
 
-module load python/3.12.1-gcc-11.3.1-7tuhjhr
+read -p "Is Ollama already set up? [y/N]: " answer
 
-# Confirm Python version & paths
-module list
-which python
-which pip
-python --version
-
-mkdir -p ~/venvs
-if [ -d "~/venvs/$venv_name" ]; then
-    rm -rf ~/venvs/$venv_name
+if [[ ! "$answer" =~ ^[yY]$ ]]; then
+    # Manual installation of Ollama
+    curl -LO https://ollama.com/download/ollama-linux-amd64.tgz
+    tar -xzf ollama-linux-amd64.tgz
+    mkdir -p $HOME/bin
+    mv ollama/bin/ollama $HOME/bin/ollama
+    echo -e "\nSuccessfully installed Ollama to '$HOME/bin/ollama'."
 fi
-python3 -m venv ~/venvs/$venv_name
-ln -s ~/venvs/$venv_name ~/work/$venv_name
-source ~/work/$venv_name/bin/activate
-
-pip install -U pip
-
-pip install inspect_ai \
-            git+https://github.com/UKGovernmentBEIS/inspect_evals \
-            vllm \
-            mistral_common # Mistral-Large-Instruct-2411
-            # openai
-
-echo -e "\nSuccessfully installed packages to '~/work/$venv_name'."
-
